requirements.txt
alembic==1.2.0
astroid==2.4.2
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.9
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
Pillow==8.1.0
pylint==2.6.0
pyparsing==2.4.6
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
toml==0.10.2
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.2.1
